+++ app/modules/blog/constants.ts
@@ -1,0 +1,3 @@
+import path from 'path';
+
+export const POSTS_DIR = path.join(import.meta.dirname, '../../../posts');
