+++ app/modules/blog/pages/PostDetailPage.tsx
@@ -2,7 +2,8 @@ import { BlocksRenderer } from '~/components/base/notion/BlocksRenderer';
 import { PostMetaData } from '~/components/base/PostMetaData';
 import { Tags } from '~/components/base/Tag';
 import { PostSwitcher } from '~/components/sections/PostSwitcher';
-import { FullPost, Post } from '~/services/post.server';
+
+import { FullPost, Post } from '../types';
 
 export interface PostDetailPageProps {
   post: FullPost;
