+++ app/modules/project/constants.ts
@@ -1,0 +1,3 @@
+import path from 'path';
+
+export const PROJECTS_FILE = path.join(import.meta.dirname, '../../../extras/projects.json');
