+++ app/routes/blog._index.tsx
@@ -1,7 +1,7 @@
 import { MetaFunction, useLoaderData } from 'react-router';
 
 import PostListPage from '~/modules/blog/pages/PostListPage';
-import { getPosts } from '~/services/post.server';
+import { getPosts } from '~/modules/blog/services/PostService/getPosts';
 
 export const meta: MetaFunction = () => {
   return [
